Renamed applist module to apps
