+++ .idea/compiler.xml
@@ -15,6 +15,7 @@
       <module name="Geto.core.screenshot-testing.test" target="17" />
       <module name="Geto.core.securesettings-test" target="17" />
       <module name="Geto.core.shortcutmanager-test" target="17" />
+      <module name="Geto.feature.applist" target="17" />
       <module name="Geto.feature.securesettingslist" target="17" />
       <module name="Geto.feature.settings.test" target="17" />
       <module name="Geto.feature.shortcut" target="17" />
