+++ .idea/gradle.xml
@@ -45,7 +45,7 @@
             <option value="$PROJECT_DIR$/core/testing" />
             <option value="$PROJECT_DIR$/core/ui" />
             <option value="$PROJECT_DIR$/feature" />
-            <option value="$PROJECT_DIR$/feature/applist" />
+            <option value="$PROJECT_DIR$/feature/apps" />
             <option value="$PROJECT_DIR$/feature/appsettings" />
             <option value="$PROJECT_DIR$/feature/settings" />
             <option value="$PROJECT_DIR$/lint" />
