+++ app-geto-catalog/dependencies/releaseRuntimeClasspath.txt
@@ -1,0 +1,90 @@
+androidx.activity:activity-compose:1.8.2
+androidx.activity:activity-ktx:1.8.2
+androidx.activity:activity:1.8.2
+androidx.annotation:annotation-experimental:1.4.0
+androidx.annotation:annotation-jvm:1.7.1
+androidx.annotation:annotation:1.7.1
+androidx.appcompat:appcompat-resources:1.6.1
+androidx.arch.core:core-common:2.2.0
+androidx.arch.core:core-runtime:2.2.0
+androidx.autofill:autofill:1.0.0
+androidx.collection:collection-jvm:1.4.0
+androidx.collection:collection-ktx:1.4.0
+androidx.collection:collection:1.4.0
+androidx.compose.animation:animation-android:1.6.4
+androidx.compose.animation:animation-core-android:1.6.4
+androidx.compose.animation:animation-core:1.6.4
+androidx.compose.animation:animation:1.6.4
+androidx.compose.foundation:foundation-android:1.6.4
+androidx.compose.foundation:foundation-layout-android:1.6.4
+androidx.compose.foundation:foundation-layout:1.6.4
+androidx.compose.foundation:foundation:1.6.4
+androidx.compose.material3:material3-android:1.2.1
+androidx.compose.material3:material3:1.2.1
+androidx.compose.material:material-icons-core-android:1.6.4
+androidx.compose.material:material-icons-core:1.6.4
+androidx.compose.material:material-icons-extended-android:1.6.4
+androidx.compose.material:material-icons-extended:1.6.4
+androidx.compose.material:material-ripple-android:1.6.4
+androidx.compose.material:material-ripple:1.6.4
+androidx.compose.runtime:runtime-android:1.6.4
+androidx.compose.runtime:runtime-saveable-android:1.6.4
+androidx.compose.runtime:runtime-saveable:1.6.4
+androidx.compose.runtime:runtime:1.6.4
+androidx.compose.ui:ui-android:1.6.4
+androidx.compose.ui:ui-geometry-android:1.6.4
+androidx.compose.ui:ui-geometry:1.6.4
+androidx.compose.ui:ui-graphics-android:1.6.4
+androidx.compose.ui:ui-graphics:1.6.4
+androidx.compose.ui:ui-text-android:1.6.4
+androidx.compose.ui:ui-text:1.6.4
+androidx.compose.ui:ui-tooling-preview-android:1.6.4
+androidx.compose.ui:ui-tooling-preview:1.6.4
+androidx.compose.ui:ui-unit-android:1.6.4
+androidx.compose.ui:ui-unit:1.6.4
+androidx.compose.ui:ui-util-android:1.6.4
+androidx.compose.ui:ui-util:1.6.4
+androidx.compose.ui:ui:1.6.4
+androidx.compose:compose-bom:2024.03.00
+androidx.concurrent:concurrent-futures:1.1.0
+androidx.core:core-ktx:1.12.0
+androidx.core:core:1.12.0
+androidx.customview:customview-poolingcontainer:1.0.0
+androidx.emoji2:emoji2:1.3.0
+androidx.exifinterface:exifinterface:1.3.7
+androidx.interpolator:interpolator:1.0.0
+androidx.lifecycle:lifecycle-common-java8:2.7.0
+androidx.lifecycle:lifecycle-common:2.7.0
+androidx.lifecycle:lifecycle-livedata-core:2.7.0
+androidx.lifecycle:lifecycle-process:2.7.0
+androidx.lifecycle:lifecycle-runtime-ktx:2.7.0
+androidx.lifecycle:lifecycle-runtime:2.7.0
+androidx.lifecycle:lifecycle-viewmodel-ktx:2.7.0
+androidx.lifecycle:lifecycle-viewmodel-savedstate:2.7.0
+androidx.lifecycle:lifecycle-viewmodel:2.7.0
+androidx.profileinstaller:profileinstaller:1.3.1
+androidx.savedstate:savedstate-ktx:1.2.1
+androidx.savedstate:savedstate:1.2.1
+androidx.startup:startup-runtime:1.1.1
+androidx.tracing:tracing:1.0.0
+androidx.vectordrawable:vectordrawable-animated:1.1.0
+androidx.vectordrawable:vectordrawable:1.1.0
+androidx.versionedparcelable:versionedparcelable:1.1.1
+com.google.accompanist:accompanist-drawablepainter:0.32.0
+com.google.guava:listenablefuture:1.0
+com.squareup.okhttp3:okhttp:4.12.0
+com.squareup.okio:okio-jvm:3.8.0
+com.squareup.okio:okio:3.8.0
+io.coil-kt:coil-base:2.6.0
+io.coil-kt:coil-compose-base:2.6.0
+io.coil-kt:coil-compose:2.6.0
+io.coil-kt:coil:2.6.0
+org.jetbrains.kotlin:kotlin-stdlib-common:1.9.22
+org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.9.0
+org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.9.0
+org.jetbrains.kotlin:kotlin-stdlib:1.9.22
+org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3
+org.jetbrains.kotlinx:kotlinx-coroutines-bom:1.7.3
+org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.7.3
+org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3
+org.jetbrains:annotations:23.0.0
