+++ benchmarks/src/main/kotlin/com/android/geto/benchmarks/appsettings/AppSettingsAction.kt
@@ -19,13 +19,13 @@ package com.android.geto.benchmarks.appsettings
 
 import androidx.benchmark.macro.MacrobenchmarkScope
 import androidx.test.uiautomator.By
-import com.android.geto.benchmarks.applist.appListClickFirstItem
+import com.android.geto.benchmarks.apps.appsClickFirstItem
 import com.android.geto.benchmarks.waitForLoadingWheelToDisappear
 import org.junit.Assert.fail
 
 fun MacrobenchmarkScope.goToAppSettingsScreen() {
     waitForLoadingWheelToDisappear()
-    appListClickFirstItem()
+    appsClickFirstItem()
     device.waitForIdle()
 
     getTopAppBar()
