--- feature/applist/src/main/kotlin/com/android/geto/feature/applist/AppListUiState.kt
@@ -1,26 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.android.geto.feature.applist
-
-import com.android.geto.core.model.TargetApplicationInfo
-
-sealed interface AppListUiState {
-    data class Success(val targetApplicationInfoList: List<TargetApplicationInfo>) : AppListUiState
-
-    data object Loading : AppListUiState
-}
