--- feature/applist/src/main/kotlin/com/android/geto/feature/applist/AppListViewModel.kt
@@ -1,43 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.android.geto.feature.applist
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.android.geto.core.data.repository.PackageRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class AppListViewModel @Inject constructor(
-    private val packageRepository: PackageRepository,
-) : ViewModel() {
-    private val _appListUiState = MutableStateFlow<AppListUiState>(AppListUiState.Loading)
-
-    val appListUiState = _appListUiState.asStateFlow()
-
-    fun getInstalledApplications() {
-        viewModelScope.launch {
-            _appListUiState.update { AppListUiState.Success(packageRepository.getInstalledApplications()) }
-        }
-    }
-}
