+++ feature/apps/src/main/AndroidManifest.xml
