+++ feature/apps/src/test/screenshots/AppsScreenLoading_foldable.png
