+++ feature/apps/src/test/screenshots/AppsScreenLoading_phone.png
