+++ feature/apps/src/test/screenshots/AppsScreenLoading_phone_dark.png
