+++ feature/apps/src/test/screenshots/AppsScreenLoading_tablet.png
