+++ feature/apps/src/test/screenshots/AppsScreenPopulated_foldable.png
