+++ feature/apps/src/test/screenshots/AppsScreenPopulated_phone.png
