+++ feature/apps/src/test/screenshots/AppsScreenPopulated_phone_dark.png
