+++ feature/apps/src/test/screenshots/AppsScreenPopulated_tablet.png
